requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-sql-utils==0.7.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-dataclasses==1.3.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
Faker==0.7.4
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
mysqlclient==2.2.4
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
pytest==8.0.2
pytest-django==4.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
split-settings==1.0.0
sqlparse==0.4.4
stripe==9.7.0
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
